requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
mss==10.1.0
pillow==12.0.0
pyperclip==1.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.18
PyYAML==6.0.3
requests==2.32.5
urllib3==2.5.0
